+++ sdk/tag.go
@@ -200,12 +200,13 @@ type RequestTagCreateTagDynamicRulesRules struct {
 }
 type RequestTagCreateTagDynamicRulesRulesItems interface{}
 type RequestTagUpdatesTagMembership struct {
-	MemberToTags *[]RequestTagUpdatesTagMembershipMemberToTags `json:"memberToTags,omitempty"` //
-	MemberType   string                                        `json:"memberType,omitempty"`   //
+	MemberToTags map[string][]string `json:"memberToTags,omitempty"` //
+	MemberType   string              `json:"memberType,omitempty"`   //
 }
-type RequestTagUpdatesTagMembershipMemberToTags struct {
-	Key []string `json:"key,omitempty"` //
-}
+
+//type RequestTagUpdatesTagMembershipMemberToTags map[string][]string {
+//	Key []string `json:"key,omitempty"` //
+//}
 type RequestTagAddMembersToTheTag map[string][]string
 
 //GetTag Get Tag - ee9a-ab01-487a-8896