+++ CHANGELOG.md
@@ -6,6 +6,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [3.5.1] - 2022-03-1
+### Added
+- Added new struct `ResponseWirelessPSKOverride`  on `wireless`for `PSKOverride` response.
+### Changed
+- Changed type of following variable on `RequestTagUpdatesTagMembership` struct of `tag`.
+    + `MemberToTags` from []RequestTagUpdatesTagMembershipMemberToTags to map[string][]string.
+### Removed
+- Removed `RequestTagUpdatesTagMembershipMemberToTags` struct from `tag`.
 
 ## [3.5.0] - 2022-02-17
 ### Added
@@ -280,4 +288,5 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 [3.4.0]: https://github.com/cisco-en-programmability/dnacenter-go-sdk/compare/v3.3.1...v3.4.0
 [3.4.1]: https://github.com/cisco-en-programmability/dnacenter-go-sdk/compare/v3.4.0...v3.4.1
 [3.5.0]: https://github.com/cisco-en-programmability/dnacenter-go-sdk/compare/v3.4.1...v3.5.0
-[Unreleased]: https://github.com/cisco-en-programmability/dnacenter-go-sdk/compare/v3.5.0...main
+[3.5.1]: https://github.com/cisco-en-programmability/dnacenter-go-sdk/compare/v3.5.0...v3.5.1
+[Unreleased]: https://github.com/cisco-en-programmability/dnacenter-go-sdk/compare/v3.5.1...main